changed build.sh

+++ build.sh
@@ -1,13 +1,17 @@
-git checkout master &&
-mkdir dontdelete  &&
-cp -r _site dontdelete/_site &&
-cp index.html dontdelete/index.html &&
+if [ -d "_site" ]; then
+  git checkout master &&
+  mkdir dontdelete  &&
+  cp -r _site dontdelete/_site &&
+  cp CNAME dontdelete/CNAME &&
 
-find . -maxdepth 1 ! -name 'dontdelete' ! -name '.*' | xargs rm -rf &&
-cp -r dontdelete/_site/* . &&
-cp -r dontdelete/index.html .
+  find . -maxdepth 1 ! -name 'dontdelete' ! -name '.*' | xargs rm -rf &&
+  cp -r dontdelete/_site/* . &&
+  cp -r dontdelete/CNAME .
 
-rm -rf dontdelete
+  rm -rf dontdelete
 
 
-touch .nojekyll
+  touch .nojekyll
+else
+  echo "Make sure to generate _site first"
+fi
